refactor: update Retrieve handling to use RetrieveWrapper
